Fix clang++-6.0 warning about different types.
pulseview/pv/data/decodesignal.cpp:590:44:
  error: comparison of integers of different signs:
  'std::deque::size_type' (aka 'unsigned long') and
  'int64_t' (aka 'long') [-Werror,-Wsign-compare]

+++ pv/data/decodesignal.cpp
@@ -575,7 +575,7 @@ void DecodeSignal::restore_settings(QSettings &settings)
 
 uint32_t DecodeSignal::get_input_segment_count() const
 {
-	int64_t count = std::numeric_limits<int64_t>::max();
+	size_t count = std::numeric_limits<int64_t>::max();
 	bool no_signals_assigned = true;
 
 	for (const data::DecodeChannel &ch : channels_)
